+++ packages/client/src/views/OfficeHome/tabs/readyForReview/ReadyForReview.tsx
@@ -9,74 +9,89 @@
  * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
-
 import {
   goToDeclarationRecordAudit,
-  goToPrintCertificate
+  goToPage,
+  goToReviewDuplicate
 } from '@client/navigation'
+import { REVIEW_EVENT_PARENT_FORM_PAGE } from '@client/navigation/routes'
+import { getScope } from '@client/profile/profileSelectors'
 import { transformData } from '@client/search/transformer'
-import { ITheme } from '@client/styledComponents'
+import { IStoreState } from '@client/store'
+import styled, { ITheme } from '@client/styledComponents'
+import { Scope, hasRegisterScope } from '@client/utils/authUtils'
 import {
   ColumnContentAlignment,
   GridTable,
   IAction,
-  SORT_ORDER,
-  COLUMNS
+  COLUMNS,
+  SORT_ORDER
 } from '@opencrvs/components/lib/interface'
 import { GQLEventSearchResultSet } from '@opencrvs/gateway/src/graphql/schema'
 import * as React from 'react'
 import { injectIntl, WrappedComponentProps as IntlShapeProps } from 'react-intl'
 import { connect } from 'react-redux'
-import { withTheme } from 'styled-components'
+import ReactTooltip from 'react-tooltip'
 import {
-  buttonMessages,
   constantsMessages,
   dynamicConstantsMessages
 } from '@client/i18n/messages'
-import { IStoreState } from '@client/store'
-import { IDeclaration, DOWNLOAD_STATUS } from '@client/declarations'
+import { messages } from '@client/i18n/messages/views/registrarHome'
+import {
+  IDeclaration,
+  DOWNLOAD_STATUS,
+  SUBMISSION_STATUS
+} from '@client/declarations'
 import { Action } from '@client/forms'
 import { DownloadButton } from '@client/components/interface/DownloadButton'
+import { withTheme } from 'styled-components'
 import { formattedDuration } from '@client/utils/date-formatting'
 import { navigationMessages } from '@client/i18n/messages/views/navigation'
-import {
-  changeSortedColumn,
-  getSortedItems
-} from '@client/views/OfficeHome/tabs/utils'
 import {
   IconWithName,
   IconWithNameEvent
 } from '@client/views/OfficeHome/tabs/components'
-import { WQContentWrapper } from '@client/views/OfficeHome/tabs/WQContentWrapper'
+import {
+  changeSortedColumn,
+  getSortedItems
+} from '@client/views/OfficeHome/tabs/utils'
 import { Downloaded } from '@opencrvs/components/lib/icons/Downloaded'
-interface IBasePrintTabProps {
+import { WQContentWrapper } from '@client/views/OfficeHome/tabs/WQContentWrapper'
+
+const ToolTipContainer = styled.span`
+  text-align: center;
+`
+interface IBaseReviewTabProps {
   theme: ITheme
-  goToPrintCertificate: typeof goToPrintCertificate
+  scope: Scope | null
+  goToPage: typeof goToPage
+  goToReviewDuplicate: typeof goToReviewDuplicate
   goToDeclarationRecordAudit: typeof goToDeclarationRecordAudit
   outboxDeclarations: IDeclaration[]
   queryData: {
     data: GQLEventSearchResultSet
   }
   paginationId: number
+  pageSize: number
   onPageChange: (newPageNumber: number) => void
   loading?: boolean
   error?: boolean
-  pageSize: number
 }
 
-interface IPrintTabState {
+interface IReviewTabState {
   width: number
   sortedCol: COLUMNS
   sortOrder: SORT_ORDER
 }
 
-type IPrintTabProps = IntlShapeProps & IBasePrintTabProps
+type IReviewTabProps = IntlShapeProps & IBaseReviewTabProps
 
-class PrintTabComponent extends React.Component<
-  IPrintTabProps,
-  IPrintTabState
+class ReadyForReviewComponent extends React.Component<
+  IReviewTabProps,
+  IReviewTabState
 > {
-  constructor(props: IPrintTabProps) {
+  pageSize = 10
+  constructor(props: IReviewTabProps) {
     super(props)
     this.state = {
       width: window.innerWidth,
@@ -97,10 +112,8 @@ class PrintTabComponent extends React.Component<
     this.setState({ width: window.innerWidth })
   }
 
-  getExpandable = () => {
-    return this.state.width > this.props.theme.grid.breakpoints.lg
-      ? true
-      : false
+  userHasRegisterScope() {
+    return this.props.scope && hasRegisterScope(this.props.scope)
   }
 
   onColumnClick = (columnName: string) => {
@@ -115,80 +128,25 @@ class PrintTabComponent extends React.Component<
     })
   }
 
-  getColumns = () => {
-    if (this.state.width > this.props.theme.grid.breakpoints.lg) {
-      return [
-        {
-          width: 30,
-          label: this.props.intl.formatMessage(constantsMessages.name),
-          key: COLUMNS.ICON_WITH_NAME,
-          isSorted: this.state.sortedCol === COLUMNS.NAME,
-          sortFunction: this.onColumnClick
-        },
-        {
-          label: this.props.intl.formatMessage(constantsMessages.event),
-          width: 16,
-          key: COLUMNS.EVENT,
-          isSorted: this.state.sortedCol === COLUMNS.EVENT,
-          sortFunction: this.onColumnClick
-        },
-        {
-          label: this.props.intl.formatMessage(constantsMessages.eventDate),
-          width: 18,
-          key: COLUMNS.DATE_OF_EVENT,
-          isSorted: this.state.sortedCol === COLUMNS.DATE_OF_EVENT,
-          sortFunction: this.onColumnClick
-        },
-        {
-          label: this.props.intl.formatMessage(constantsMessages.registered),
-          width: 18,
-          key: COLUMNS.REGISTERED,
-          isSorted: this.state.sortedCol === COLUMNS.REGISTERED,
-          sortFunction: this.onColumnClick
-        },
-        {
-          width: 18,
-          alignment: ColumnContentAlignment.RIGHT,
-          key: COLUMNS.ACTIONS,
-          isActionColumn: true
-        }
-      ]
-    } else {
-      return [
-        {
-          label: this.props.intl.formatMessage(constantsMessages.name),
-          width: 70,
-          key: COLUMNS.ICON_WITH_NAME_EVENT
-        },
-        {
-          width: 30,
-          alignment: ColumnContentAlignment.RIGHT,
-          key: COLUMNS.ACTIONS,
-          isActionColumn: true
-        }
-      ]
-    }
-  }
-
-  transformRegisteredContent = (data: GQLEventSearchResultSet) => {
+  transformDeclaredContent = (data: GQLEventSearchResultSet) => {
     const { intl } = this.props
     if (!data || !data.results) {
       return []
     }
-
     const transformedData = transformData(data, this.props.intl)
     const items = transformedData.map((reg, index) => {
+      const actions = [] as IAction[]
       const foundDeclaration = this.props.outboxDeclarations.find(
         (declaration) => declaration.id === reg.id
       )
-      const actions: IAction[] = []
       const downloadStatus =
         (foundDeclaration && foundDeclaration.downloadStatus) || undefined
+      const isDuplicate = reg.duplicates && reg.duplicates.length > 0
 
       if (downloadStatus !== DOWNLOAD_STATUS.DOWNLOADED) {
         if (this.state.width > this.props.theme.grid.breakpoints.lg) {
           actions.push({
-            label: this.props.intl.formatMessage(buttonMessages.print),
+            label: this.props.intl.formatMessage(constantsMessages.review),
             handler: () => {},
             disabled: true
           })
@@ -209,19 +167,25 @@ class PrintTabComponent extends React.Component<
       } else {
         if (this.state.width > this.props.theme.grid.breakpoints.lg) {
           actions.push({
-            label: this.props.intl.formatMessage(buttonMessages.print),
+            label: this.props.intl.formatMessage(constantsMessages.review),
             handler: (
               e: React.MouseEvent<HTMLButtonElement, MouseEvent> | undefined
             ) => {
               e && e.stopPropagation()
-              this.props.goToPrintCertificate(
-                reg.id,
-                reg.event.toLocaleLowerCase() || ''
-              )
+              isDuplicate
+                ? this.props.goToReviewDuplicate(reg.id)
+                : this.props.goToPage(
+                    REVIEW_EVENT_PARENT_FORM_PAGE,
+                    reg.id,
+                    'review',
+                    reg.event ? reg.event.toLowerCase() : ''
+                  )
             }
           })
         }
-        actions.push({ actionComponent: <Downloaded /> })
+        actions.push({
+          actionComponent: <Downloaded />
+        })
       }
       const event =
         (reg.event &&
@@ -229,36 +193,47 @@ class PrintTabComponent extends React.Component<
             dynamicConstantsMessages[reg.event.toLowerCase()]
           )) ||
         ''
+      const isValidatedOnReview =
+        reg.declarationStatus === SUBMISSION_STATUS.VALIDATED &&
+        this.userHasRegisterScope()
+          ? true
+          : false
       const dateOfEvent =
-        reg.dateOfEvent &&
-        reg.dateOfEvent.length > 0 &&
-        new Date(reg.dateOfEvent)
-      const registered =
-        (reg.modifiedAt && Number.isNaN(Number(reg.modifiedAt))
-          ? new Date(reg.modifiedAt)
-          : new Date(Number(reg.modifiedAt))) || ''
+        (reg.dateOfEvent &&
+          reg.dateOfEvent.length > 0 &&
+          new Date(reg.dateOfEvent)) ||
+        ''
+      const createdAt = (reg.createdAt && parseInt(reg.createdAt)) || ''
+      console.log(reg)
       return {
         ...reg,
         event,
+        dateOfEvent,
+        sentForReview: createdAt,
         name: reg.name && reg.name.toLowerCase(),
         iconWithName: (
-          <IconWithName status={reg.declarationStatus} name={reg.name} />
+          <IconWithName
+            status={reg.declarationStatus}
+            name={reg.name}
+            isDuplicate={isDuplicate}
+            isValidatedOnReview={isValidatedOnReview}
+          />
         ),
         iconWithNameEvent: (
           <IconWithNameEvent
             status={reg.declarationStatus}
             name={reg.name}
             event={event}
+            isValidatedOnReview={isValidatedOnReview}
+            isDuplicate={isDuplicate}
           />
         ),
-        dateOfEvent,
-        registered,
         actions,
         rowClickHandler: [
           {
             label: 'rowClickHandler',
             handler: () =>
-              this.props.goToDeclarationRecordAudit('printTab', reg.id)
+              this.props.goToDeclarationRecordAudit('reviewTab', reg.id)
           }
         ]
       }
@@ -268,20 +243,74 @@ class PrintTabComponent extends React.Component<
       this.state.sortedCol,
       this.state.sortOrder
     )
-
     return sortedItems.map((item) => {
       return {
         ...item,
         dateOfEvent:
           item.dateOfEvent && formattedDuration(item.dateOfEvent as Date),
-        registered:
-          item.registered && formattedDuration(item.registered as Date)
+        sentForReview:
+          item.sentForReview && formattedDuration(item.sentForReview as number)
       }
     })
   }
 
+  getColumns = () => {
+    if (this.state.width > this.props.theme.grid.breakpoints.lg) {
+      return [
+        {
+          label: this.props.intl.formatMessage(constantsMessages.name),
+          width: 30,
+          key: COLUMNS.ICON_WITH_NAME,
+          sortFunction: this.onColumnClick,
+          isSorted: this.state.sortedCol === COLUMNS.NAME
+        },
+        {
+          label: this.props.intl.formatMessage(constantsMessages.name),
+          width: 16,
+          key: COLUMNS.EVENT,
+          sortFunction: this.onColumnClick,
+          isSorted: this.state.sortedCol === COLUMNS.EVENT
+        },
+        {
+          label: this.props.intl.formatMessage(constantsMessages.eventDate),
+          width: 18,
+          key: COLUMNS.DATE_OF_EVENT,
+          sortFunction: this.onColumnClick,
+          isSorted: this.state.sortedCol === COLUMNS.DATE_OF_EVENT
+        },
+        {
+          label: this.props.intl.formatMessage(constantsMessages.sentForReview),
+          width: 18,
+          key: COLUMNS.SENT_FOR_REVIEW,
+          sortFunction: this.onColumnClick,
+          isSorted: this.state.sortedCol === COLUMNS.SENT_FOR_REVIEW
+        },
+        {
+          width: 18,
+          key: COLUMNS.ACTIONS,
+          isActionColumn: true,
+          alignment: ColumnContentAlignment.RIGHT
+        }
+      ]
+    } else {
+      return [
+        {
+          label: this.props.intl.formatMessage(constantsMessages.name),
+          width: 70,
+          key: COLUMNS.ICON_WITH_NAME_EVENT
+        },
+        {
+          width: 30,
+          alignment: ColumnContentAlignment.RIGHT,
+          key: COLUMNS.ACTIONS,
+          isActionColumn: true
+        }
+      ]
+    }
+  }
+
   render() {
-    const { intl, queryData, paginationId, onPageChange, pageSize } = this.props
+    const { intl, queryData, paginationId, pageSize, onPageChange } = this.props
     const { data } = queryData
     const totalPages = this.props.queryData.data.totalItems
       ? Math.ceil(this.props.queryData.data.totalItems / pageSize)
@@ -293,8 +322,10 @@ class PrintTabComponent extends React.Component<
         : false
     return (
       <WQContentWrapper
-        title={intl.formatMessage(navigationMessages.print)}
-        isMobileSize={this.state.width < this.props.theme.grid.breakpoints.lg}
+        title={intl.formatMessage(navigationMessages.readyForReview)}
+        isMobileSize={
+          this.state.width < this.props.theme.grid.breakpoints.lg ? true : false
+        }
         isShowPagination={isShowPagination}
         paginationId={paginationId}
         totalPages={totalPages}
@@ -302,12 +333,19 @@ class PrintTabComponent extends React.Component<
         loading={this.props.loading}
         error={this.props.error}
         noResultText={intl.formatMessage(constantsMessages.noRecords, {
-          tab: 'ready to print'
+          tab: 'are ready for review'
         })}
-        noContent={this.transformRegisteredContent(data).length <= 0}
+        noContent={this.transformDeclaredContent(data).length <= 0}
       >
+        <ReactTooltip id="validateTooltip">
+          <ToolTipContainer>
+            {this.props.intl.formatMessage(
+              messages.validatedDeclarationTooltipForRegistrar
+            )}
+          </ToolTipContainer>
+        </ReactTooltip>
         <GridTable
-          content={this.transformRegisteredContent(data)}
+          content={this.transformDeclaredContent(data)}
           columns={this.getColumns()}
           clickable={true}
           loading={this.props.loading}
@@ -321,11 +359,13 @@ class PrintTabComponent extends React.Component<
 
 function mapStateToProps(state: IStoreState) {
   return {
+    scope: getScope(state),
     outboxDeclarations: state.declarationsState.declarations
   }
 }
 
-export const PrintTab = connect(mapStateToProps, {
-  goToPrintCertificate,
+export const ReadyForReview = connect(mapStateToProps, {
+  goToPage,
+  goToReviewDuplicate,
   goToDeclarationRecordAudit
-})(injectIntl(withTheme(PrintTabComponent)))
+})(injectIntl(withTheme(ReadyForReviewComponent)))
