+++ packages/client/src/views/OfficeHome/tabs/sentForApproval/SentForApproval.tsx
@@ -9,21 +9,19 @@
  * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
+import { IDeclaration } from '@client/declarations'
 import {
-  goToDeclarationRecordAudit,
-  goToPage,
-  goToReviewDuplicate
-} from '@client/navigation'
-import { REVIEW_EVENT_PARENT_FORM_PAGE } from '@client/navigation/routes'
+  constantsMessages,
+  dynamicConstantsMessages
+} from '@client/i18n/messages'
+import { messages } from '@client/i18n/messages/views/registrarHome'
+import { goToDeclarationRecordAudit, goToPage } from '@client/navigation'
 import { getScope } from '@client/profile/profileSelectors'
 import { transformData } from '@client/search/transformer'
 import { IStoreState } from '@client/store'
-import { ITheme } from '@client/styledComponents'
-import { Scope } from '@client/utils/authUtils'
+import styled, { ITheme } from '@client/styledComponents'
 import {
-  ColumnContentAlignment,
   GridTable,
-  IAction,
   COLUMNS,
   SORT_ORDER
 } from '@opencrvs/components/lib/interface'
@@ -31,15 +29,8 @@ import { GQLEventSearchResultSet } from '@opencrvs/gateway/src/graphql/schema'
 import * as React from 'react'
 import { injectIntl, WrappedComponentProps as IntlShapeProps } from 'react-intl'
 import { connect } from 'react-redux'
+import ReactTooltip from 'react-tooltip'
 import { withTheme } from 'styled-components'
-import {
-  buttonMessages,
-  constantsMessages,
-  dynamicConstantsMessages
-} from '@client/i18n/messages'
-import { IDeclaration, DOWNLOAD_STATUS } from '@client/declarations'
-import { Action } from '@client/forms'
-import { DownloadButton } from '@client/components/interface/DownloadButton'
 import { formattedDuration } from '@client/utils/date-formatting'
 import { navigationMessages } from '@client/i18n/messages/views/navigation'
 import {
@@ -51,13 +42,12 @@ import {
   IconWithNameEvent
 } from '@client/views/OfficeHome/tabs/components'
 import { WQContentWrapper } from '@client/views/OfficeHome/tabs/WQContentWrapper'
-import { Downloaded } from '@opencrvs/components/lib/icons/Downloaded'
-
-interface IBaseRejectTabProps {
+const ToolTipContainer = styled.span`
+  text-align: center;
+`
+interface IBaseApprovalTabProps {
   theme: ITheme
-  scope: Scope | null
   goToPage: typeof goToPage
-  goToReviewDuplicate: typeof goToReviewDuplicate
   goToDeclarationRecordAudit: typeof goToDeclarationRecordAudit
   outboxDeclarations: IDeclaration[]
   queryData: {
@@ -70,29 +60,26 @@ interface IBaseRejectTabProps {
   error?: boolean
 }
 
-interface IRejectTabState {
+interface IApprovalTabState {
   width: number
   sortedCol: COLUMNS
   sortOrder: SORT_ORDER
 }
 
-type IRejectTabProps = IntlShapeProps & IBaseRejectTabProps
+type IApprovalTabProps = IntlShapeProps & IBaseApprovalTabProps
 
-class RejectTabComponent extends React.Component<
-  IRejectTabProps,
-  IRejectTabState
+class SentForApprovalComponent extends React.Component<
+  IApprovalTabProps,
+  IApprovalTabState
 > {
-  constructor(props: IRejectTabProps) {
+  pageSize = 10
+  constructor(props: IApprovalTabProps) {
     super(props)
     this.state = {
       width: window.innerWidth,
       sortedCol: COLUMNS.NAME,
       sortOrder: SORT_ORDER.ASCENDING
     }
-  }
-
-  userHasRegisterScope() {
-    return this.props.scope && this.props.scope.includes('register')
   }
 
   componentDidMount() {
@@ -145,103 +132,40 @@ class RejectTabComponent extends React.Component<
         },
         {
           label: this.props.intl.formatMessage(
-            constantsMessages.sentForUpdates
+            constantsMessages.sentForApproval
           ),
           width: 18,
-          key: COLUMNS.SENT_FOR_UPDATES,
-          isSorted: this.state.sortedCol === COLUMNS.SENT_FOR_UPDATES,
+          key: COLUMNS.SENT_FOR_APPROVAL,
+          isSorted: this.state.sortedCol === COLUMNS.SENT_FOR_APPROVAL,
           sortFunction: this.onColumnClick
-        },
-        {
-          width: 18,
-          alignment: ColumnContentAlignment.RIGHT,
-          key: COLUMNS.ACTIONS,
-          isActionColumn: true
         }
       ]
     } else {
       return [
         {
           label: this.props.intl.formatMessage(constantsMessages.name),
-          width: 70,
+          width: 90,
           key: COLUMNS.ICON_WITH_NAME_EVENT
-        },
-        {
-          width: 30,
-          alignment: ColumnContentAlignment.RIGHT,
-          key: COLUMNS.ACTIONS,
-          isActionColumn: true
         }
       ]
     }
   }
 
-  transformRejectedContent = (data: GQLEventSearchResultSet) => {
+  transformValidatedContent = (data: GQLEventSearchResultSet) => {
     const { intl } = this.props
     if (!data || !data.results) {
       return []
     }
     const transformedData = transformData(data, this.props.intl)
-    const items = transformedData.map((reg, index) => {
-      const actions = [] as IAction[]
-      const foundDeclaration = this.props.outboxDeclarations.find(
-        (declaration) => declaration.id === reg.id
-      )
-      const downloadStatus =
-        (foundDeclaration && foundDeclaration.downloadStatus) || undefined
-      const isDuplicate = reg.duplicates && reg.duplicates.length > 0
 
-      if (downloadStatus !== DOWNLOAD_STATUS.DOWNLOADED) {
-        if (this.state.width > this.props.theme.grid.breakpoints.lg) {
-          actions.push({
-            label: this.props.intl.formatMessage(buttonMessages.update),
-            handler: () => {},
-            disabled: true
-          })
-        }
-        actions.push({
-          actionComponent: (
-            <DownloadButton
-              downloadConfigs={{
-                event: reg.event,
-                compositionId: reg.id,
-                action: Action.LOAD_REVIEW_DECLARATION
-              }}
-              key={`DownloadButton-${index}`}
-              status={downloadStatus as DOWNLOAD_STATUS}
-            />
-          )
-        })
-      } else {
-        if (this.state.width > this.props.theme.grid.breakpoints.lg) {
-          actions.push({
-            label: this.props.intl.formatMessage(buttonMessages.update),
-            handler: (
-              e: React.MouseEvent<HTMLButtonElement, MouseEvent> | undefined
-            ) => {
-              e && e.stopPropagation()
-              isDuplicate
-                ? this.props.goToReviewDuplicate(reg.id)
-                : this.props.goToPage(
-                    REVIEW_EVENT_PARENT_FORM_PAGE,
-                    reg.id,
-                    'review',
-                    reg.event ? reg.event.toLowerCase() : ''
-                  )
-            }
-          })
-        }
-        actions.push({
-          actionComponent: <Downloaded />
-        })
-      }
+    const items = transformedData.map((reg) => {
       const event =
         (reg.event &&
           intl.formatMessage(
             dynamicConstantsMessages[reg.event.toLowerCase()]
           )) ||
         ''
-      const sentForUpdates =
+      const sentForApproval =
         (reg.modifiedAt && Number.isNaN(Number(reg.modifiedAt))
           ? new Date(reg.modifiedAt)
           : new Date(Number(reg.modifiedAt))) || ''
@@ -254,28 +178,25 @@ class RejectTabComponent extends React.Component<
         event,
         name: reg.name && reg.name.toLowerCase(),
         iconWithName: (
-          <IconWithName
-            status={reg.declarationStatus}
-            name={reg.name}
-            isDuplicate={isDuplicate}
-          />
+          <IconWithName status={reg.declarationStatus} name={reg.name} />
         ),
         iconWithNameEvent: (
           <IconWithNameEvent
             status={reg.declarationStatus}
             name={reg.name}
-            event={reg.event}
-            isDuplicate={isDuplicate}
+            event={event}
           />
         ),
-        sentForUpdates,
+        eventTimeElapsed:
+          (reg.dateOfEvent && formattedDuration(new Date(reg.dateOfEvent))) ||
+          '',
         dateOfEvent,
-        actions,
+        sentForApproval,
         rowClickHandler: [
           {
             label: 'rowClickHandler',
             handler: () =>
-              this.props.goToDeclarationRecordAudit('rejectTab', reg.id)
+              this.props.goToDeclarationRecordAudit('approvalTab', reg.id)
           }
         ]
       }
@@ -290,14 +211,15 @@ class RejectTabComponent extends React.Component<
         ...item,
         dateOfEvent:
           item.dateOfEvent && formattedDuration(item.dateOfEvent as Date),
-        sentForUpdates:
-          item.sentForUpdates && formattedDuration(item.sentForUpdates as Date)
+        sentForApproval:
+          item.sentForApproval &&
+          formattedDuration(item.sentForApproval as Date)
       }
     })
   }
 
   render() {
-    const { intl, queryData, paginationId, onPageChange, pageSize } = this.props
+    const { intl, queryData, paginationId, pageSize, onPageChange } = this.props
     const { data } = queryData
     const totalPages = this.props.queryData.data.totalItems
       ? Math.ceil(this.props.queryData.data.totalItems / pageSize)
@@ -309,21 +231,28 @@ class RejectTabComponent extends React.Component<
         : false
     return (
       <WQContentWrapper
-        title={intl.formatMessage(navigationMessages.sentForUpdates)}
+        title={intl.formatMessage(navigationMessages.approvals)}
         isMobileSize={this.state.width < this.props.theme.grid.breakpoints.lg}
         isShowPagination={isShowPagination}
         paginationId={paginationId}
         totalPages={totalPages}
         onPageChange={onPageChange}
+        noResultText={intl.formatMessage(constantsMessages.noRecords, {
+          tab: 'sent for approval'
+        })}
         loading={this.props.loading}
         error={this.props.error}
-        noResultText={intl.formatMessage(constantsMessages.noRecords, {
-          tab: 'sent for updates'
-        })}
-        noContent={this.transformRejectedContent(data).length <= 0}
+        noContent={this.transformValidatedContent(data).length <= 0}
       >
+        <ReactTooltip id="validatedTooltip">
+          <ToolTipContainer>
+            {this.props.intl.formatMessage(
+              messages.validatedDeclarationTooltipForRegistrationAgent
+            )}
+          </ToolTipContainer>
+        </ReactTooltip>
         <GridTable
-          content={this.transformRejectedContent(data)}
+          content={this.transformValidatedContent(data)}
           columns={this.getColumns()}
           clickable={true}
           loading={this.props.loading}
@@ -342,8 +271,7 @@ function mapStateToProps(state: IStoreState) {
   }
 }
 
-export const RejectTab = connect(mapStateToProps, {
+export const SentForApproval = connect(mapStateToProps, {
   goToPage,
-  goToReviewDuplicate,
   goToDeclarationRecordAudit
-})(injectIntl(withTheme(RejectTabComponent)))
+})(injectIntl(withTheme(SentForApprovalComponent)))
